@@ -9,10 +9,10 @@
     "description": "CMS plugin for Sylius applications.",
     "license": "MIT",
     "require": {
-        "php": "^7.3 || ^7.4 || ^8.0",
-        "sylius/sylius": "~1.8.0 || ~1.9.0 || ~1.10.0 || ~1.11.0",
+        "php": "^8.0",
+        "sylius/sylius": "~1.11.0 || ~1.12.0",
         "league/csv": "^9.1",
-        "friendsofsymfony/ckeditor-bundle": "^1.1||^2.0",
+        "friendsofsymfony/ckeditor-bundle": "^1.1 || ^2.0",
         "instaclick/php-webdriver": "^1.4",
         "symfony/webpack-encore-bundle": "^1.12"
     },
@@ -30,8 +30,7 @@
         "friends-of-behat/suite-settings-extension": "^1.0",
         "friends-of-behat/symfony-extension": "^2.1",
         "friends-of-behat/variadic-extension": "^1.3",
-        "lchrusciel/api-test-case": "^4.1||^5.0",
-        "friendsofsymfony/oauth-server-bundle": "^1.6 || >2.0.0-alpha.0 ^2.0@dev",
+        "lchrusciel/api-test-case": "^4.1 || ^5.0",
         "phpspec/phpspec": "^7.0",
         "phpstan/extension-installer": "^1.0",
         "phpstan/phpstan": "0.12.99",
@@ -39,16 +38,17 @@
         "phpstan/phpstan-strict-rules": "^0.12.0",
         "phpstan/phpstan-webmozart-assert": "0.12.12",
         "phpunit/phpunit": "^9.5",
-        "sensiolabs/security-checker": "^6.0",
         "sylius-labs/coding-standard": "^4.0",
-        "symfony/browser-kit": "^4.4 || ^5.2",
-        "symfony/debug-bundle": "^4.4 || ^5.2",
-        "symfony/dotenv": "^4.4 || ^5.2",
-        "symfony/intl": "^4.4 || ^5.2",
-        "symfony/web-profiler-bundle": "^4.4 || ^5.2",
-        "vimeo/psalm": "4.7.1",
+        "symfony/browser-kit": "^5.4 || ^6.0",
+        "symfony/debug-bundle": "^5.4 || ^6.0",
+        "symfony/dotenv": "^5.4 || ^6.0",
+        "symfony/intl": "^5.4 || ^6.0",
+        "symfony/web-profiler-bundle": "^5.4 || ^6.0",
+        "league/flysystem-bundle": "2.4.0",
+        "sylius/mailer-bundle":  "^1.8 || ^2.0@beta",
         "matthiasnoback/symfony-config-test": "^4.3",
-        "polishsymfonycommunity/symfony-mocker-container": "^1.0"
+        "polishsymfonycommunity/symfony-mocker-container": "^1.0",
+        "twig/twig": "^2.10.0"
     },
     "conflict": {
         "symfony/symfony": "4.1.8",
